requirements.txt
PyQt6>=6.5.0
Pillow>=10.0.0
opencv-python>=4.8.0
